requirements.txt
apache-airflow-providers-apache-spark
requests
boto3
protobuf==3.20.*
cryptography==41.0.5
pyOpenSSL==24.0